+++ src/types/products.ts
@@ -1,0 +1,14 @@
+/* eslint-disable import/no-cycle */
+import { ITicketLine } from './tickets';
+
+export interface IProduct {
+  id: number;
+  codigo: string;
+  descripcion: string;
+  precio: number;
+  precioTarjeta: number;
+}
+export interface IReturnProduct {
+  line: ITicketLine;
+  returnAmount: number;
+}