+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.6.0] - 2025-06-03
+
+## Changed
+
+- Refactor types organization of the code to use a more modular approach.
+- Refactor useTicketState and useHomeState to be React Providers instead of hooks.
+- Refactor useTicketApi & useAuthApi to use TicketsAPI & AuthAPI classes.
+
 ## [1.5.0] - 2025-05-18
 
 ## Add